chore: refactor tests to hopefully pass
