+++ .github/workflows/ci.yml
@@ -28,11 +28,19 @@ jobs:
 
       - name: install plenary.nvim
         run: |
-            git clone --depth 1 https://github.com/nvim-lua/plenary.nvim ~/.local/share/nvim/site/pack/vendor/start/plenary.nvim
+            git clone --depth 1 https://github.com/nvim-lua/plenary.nvim ~/.local/share/nvim/site/pack/vendor/start/plenary
+            # git clone --depth 1 https://github.com/nvim-lua/plenary.nvim ~/.local/share/nvim/site/pack/vendor/start/plenary.nvim
 
-      - name: run test
+      - name: setup tools
         shell: bash
         run: |
           luarocks install luacheck
           luarocks install vusted
-          vusted
+
+      - name: Run tests
+        shell: bash
+        env:
+          VUSTED_NVIM: ${{ steps.vim.outputs.executable }}
+          VUSTED_ARGS: "--headless"
+        run: |
+            make test
