+++ lua/git-worktree/config_spec.lua
