+++ test/fixtures/.repo/.git-orig/COMMIT_EDITMSG
