+++ test/fixtures/.repo/.git-orig/hooks/applypatch-msg.sample
