+++ test/fixtures/.repo/.git-orig/hooks/commit-msg.sample
