+++ test/fixtures/.repo/.git-orig/hooks/pre-applypatch.sample
