+++ test/fixtures/.repo/.git-orig/hooks/pre-commit.sample
