+++ test/fixtures/.repo/.git-orig/hooks/pre-merge-commit.sample
