+++ test/fixtures/.repo/.git-orig/hooks/pre-push.sample
