+++ test/fixtures/.repo/.git-orig/hooks/pre-rebase.sample
