+++ test/fixtures/.repo/.git-orig/hooks/pre-receive.sample
