+++ test/fixtures/.repo/.git-orig/hooks/prepare-commit-msg.sample
