+++ test/fixtures/.repo/.git-orig/hooks/update.sample
